services/order_service/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
SQLAlchemy==2.0.37
psycopg2-binary==2.9.6
kafka-python==2.0.2
requests==2.28.2
python-dotenv==1.0.0
gunicorn==20.1.0
pytest==7.3.1
black==23.3.0
isort==5.12.0
mypy==1.2.0
Werkzeug==2.2.3
